@@ -13,31 +13,31 @@
   "private": true,
   "dependencies": {
     "@agm/core": "~1.0.0-beta.2",
-    "@angular/animations": "^5.2.9",
-    "@angular/cdk": "^5.2.4",
-    "@angular/common": "^5.2.6",
-    "@angular/compiler": "^5.2.6",
-    "@angular/core": "^5.2.6",
-    "@angular/forms": "^5.2.6",
-    "@angular/http": "^5.2.0",
-    "@angular/material": "^5.2.4",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "angular-bootstrap-md": "^5.2.3",
+    "@angular/animations": "7.1.4",
+    "@angular/cdk": "^7.2.0",
+    "@angular/common": "7.1.4",
+    "@angular/compiler": "7.1.4",
+    "@angular/core": "7.1.4",
+    "@angular/forms": "7.1.4",
+    "@angular/http": "7.1.4",
+    "@angular/material": "^7.2.0",
+    "@angular/platform-browser": "7.1.4",
+    "@angular/platform-browser-dynamic": "7.1.4",
+    "@angular/router": "7.1.4",
     "chart.js": "^2.5.0",
     "core-js": "^2.4.1",
-    "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
     "hoek": "5.0.3",
     "popper": "^1.0.1",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19"
+    "rxjs": "^6.3.3",
+    "tslib": "^1.9.0",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "~1.7.3",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular-devkit/build-angular": "~0.11.0",
+    "@angular/cli": "7.1.4",
+    "@angular/compiler-cli": "7.1.4",
+    "@angular/language-service": "7.1.4",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
@@ -53,6 +53,6 @@
     "protractor": "^5.4.0",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "3.1.6"
   }
 }
