@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import {Piece} from '../classes/Piece';
 import {MusicService} from './music.service';
-import {forkJoin} from 'rxjs/observable/forkJoin';
+import {forkJoin} from 'rxjs';
 import {Proposal} from '../classes/Proposal';
 import {stringify} from 'querystring';
 
